Z_CallocLevSpec i32 -> u16

+++ z_zone.h
@@ -37,7 +37,7 @@ void __far* Z_MallocStatic(uint16_t size);
 void __far* Z_MallocStaticWithUser(uint16_t size, void __far*__far* user); 
 void __far* Z_MallocLevel(uint16_t size, void __far*__far* user);
 void __far* Z_CallocLevel(uint16_t size);
-void __far* Z_CallocLevSpec(int32_t size);
+void __far* Z_CallocLevSpec(uint16_t size);
 void Z_ChangeTagToStatic(const void __far* ptr);
 void Z_ChangeTagToCache(const void __far* ptr);
 void Z_Free(const void __far* ptr);
